removed outdated font scripts, updated main menu ui to new fonts
