+++ scripts/ui_elements/frame.py
@@ -10,6 +10,7 @@ from scripts.core.constants import DEFAULT_IMAGE_SIZE, GAP_SIZE
 from scripts.core.utility import clamp
 from scripts.ui_elements.fancy_font import FancyFont
 from scripts.ui_elements.font import Font
+from scripts.ui_elements.font import Font
 
 if TYPE_CHECKING:
     from typing import List, Optional, Tuple, Union
@@ -23,19 +24,19 @@ class Frame(UIElement):
         self,
         pos: Tuple[int, int],
         image: Optional[pygame.surface] = None,
-        text_and_font: Optional[Tuple[str, Union[Font, FancyFont]]] = (None, None),
+        font: Optional[Union[Font, FancyFont]] = None,
         is_selectable: bool = False,
-        max_line_width: int = 0,
+        max_width: int = 0,
         max_height: Optional[int] = None,
     ):
         super().__init__(pos, is_selectable)
 
         self.image: Optional[pygame.surface] = image
-        self.text: Optional[str] = str(text_and_font[0])
-        self.font: Optional[Union[Font, FancyFont]] = text_and_font[1]
-        self.line_width = max_line_width
+        self.font: Optional[Union[Font, FancyFont]] = font
+        self.max_width = max_width
         self.max_height = max_height
 
+        self._override_font_attrs()
         self._rebuild_surface()
 
     def update(self, delta_time: float):
@@ -56,7 +57,6 @@ class Frame(UIElement):
 
     def _recalculate_size(self):
         image = self.image
-        text = self.text
         font = self.font
 
         width = 0
@@ -66,22 +66,23 @@ class Frame(UIElement):
             width += image.get_width()
             height += image.get_height()
 
-        if text is not None and font is not None:
-            if isinstance(font, Font):
-                width += font.width(text) + GAP_SIZE
-            else:
-                width += font.width + GAP_SIZE
-            # N.B. doesnt amend height as is drawn next to image
+        if font is not None:
+            width += font.width + GAP_SIZE
+
+            # N.B. doesnt amend height if there is an image as is drawn next to the image
             if height == 0:
                 height += self.font.height
 
         # respect max height
         if self.max_height is not None:
-            height_ = min(self.max_height, height)
+            height = min(height, self.max_height)
         else:
-            height_ = height
+            height = height
 
-        self.size = (width, height_)
+        # respect max width
+        width = min(width, self.max_width)
+
+        self.size = (width, height)
         self.surface = pygame.Surface(self.size, SRCALPHA)
 
     def _rebuild_surface(self):
@@ -89,7 +90,6 @@ class Frame(UIElement):
 
         surface = self.surface
         image = self.image
-        text = self.text
         font = self.font
 
         # draw image
@@ -97,16 +97,53 @@ class Frame(UIElement):
             surface.blit(image, (0, 0))
 
         # draw text
-        if text and font:
-            if image:
-                start_x = image.get_width() + GAP_SIZE
-            else:
-                start_x = 0
+        if font:
+            # Font can be drawn once (FancyFont needs constant redrawing)
+            if isinstance(font, Font):
+                font.render(surface)
 
-            font.render(text, surface, (start_x, font.line_height // 2), self.line_width)
+    def _override_font_attrs(self):
+        """
+        Force the font to use the frame's pos and max sizes.
+        """
+        image = self.image
+        font = self.font
+
+        if not font:
+            return
+
+        # offset for image, if there is one
+        if image:
+            image_width = image.get_width()
+            x = image_width + GAP_SIZE
+        else:
+            image_width = 0
+            x = 0
+
+        # update font pos (remember, this is relative to the frame)
+        y = font.line_height // 2
+        font.pos = (x, y)
+
+        # update font size
+        font.line_width = self.max_width - image_width
+
+        # FancyFont needs to refresh
+        if isinstance(font, FancyFont):
+            font.refresh()
+            pass
 
     def set_text(self, text: str):
-        self.text = str(text)
+        """
+        Update the font text.
+        """
+        font = self.font
+
+        if isinstance(font, FancyFont):
+            font.raw_text = text
+            font.refresh()
+
+        elif isinstance(font, Font):
+            font.text = text
 
         self._rebuild_surface()
